% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_counts.R
\name{counts_to_array}
\alias{counts_to_array}
\title{convert a raw counts table to a three-dimensional array}
\usage{
counts_to_array(raw_counts, sample_mapping, ordering)
}
\arguments{
\item{raw_counts}{the data frame of the raw counts}

\item{sample_mapping}{a data frame with the following column names:
\itemize{
\item c_name: the name of the column as it is represented in raw_counts
\item sample: a unique identifier for the sample (replicate)
\item bin: the bin that this column refers to. must be consistent with \code{ordering}.
}}

\item{ordering}{an ordering corresponding to the bin column in \code{sample_mapping}.
e.g.: c('low', 'low_mid', 'mid_high', 'high').}
}
\value{
a three-dimensional array corresponding to \link{sample, guide, bins}
}
\description{
takes a two-dimensional data frame with guides on the rows and samples + bins on the
columns and converts it to a three-dimensional array \link{sample, guide, bins}.
}
