% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nimble.R
\name{cutoff_to_probability}
\alias{cutoff_to_probability}
\title{convert standard normal cutoffs to get the corresponding mass in each bin}
\usage{
cutoff_to_probability(cutoff, offset)
}
\description{
given a set of cutoffs, convert these bins to a probability mass function.
}
