% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nimble.R
\name{ddirchmulti}
\alias{ddirchmulti}
\title{dirichlet multinomial likelihood}
\usage{
ddirchmulti(x, alpha, size, log = 0)
}
\arguments{
\item{x}{a vector of data}

\item{alpha}{a vector of concentration parameters}

\item{size}{the sum of x.}

\item{log}{if TRUE return the log probability. otherwise, just return probability.}
}
\description{
used for MCMC. slightly modified version from the NIMBLE manual.
}
