% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nimble.R
\name{dirichlet_to_normal_bins}
\alias{dirichlet_to_normal_bins}
\title{convert a N-dimensional dirichlet to a N-1-dimensional set of normal cutoffs}
\usage{
dirichlet_to_normal_bins(alpha)
}
\description{
match the quantiles from the Dirichlet to a standard normal and return the corresponding
cutoffs to match the bins
}
