% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summaries.R
\name{wb_get_posterior}
\alias{wb_get_posterior}
\title{Get the posterior of several parameters}
\usage{
wb_get_posterior(wo, wb_samples, parameter)
}
\arguments{
\item{wo}{a fully initialized waterbear object}

\item{wb_samples}{a sampled model}

\item{parameter}{the parameter that you wish to extract: 'guide', 'gene_pip', or 'gene'.}
}
\value{
a matrix where the rows represent the parameter and of the columns represent a specific sample. Note that the samples will always be in the same order, so if you want to compute functions of the samples, you can simply do the appropriate element wise operations. An example might be getting the guide level effect while incorporating the posterior inclusion probability.
}
\description{
After the sampling has been done using NIMBLE, you can extract several of the relevant parameters.
}
