% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summaries.R
\name{wb_guide_summaries}
\alias{wb_guide_summaries}
\title{Get guide level summaries}
\usage{
wb_guide_summaries(wo, wb_samples)
}
\arguments{
\item{wo}{a fully initialized waterbear object.}

\item{wb_samples}{NIMBLE samples in waterbear format}
}
\value{
a list with three components: (1) guides, (2) gene_inclusion estimates, and (3) gene_shift estimates
}
\description{
This functigon provids both guide and gene level summaries.
}
