% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_counts.R
\name{wb_make_object}
\alias{wb_make_object}
\title{counts to waterbear object}
\usage{
wb_make_object(
  counts_array,
  gene_mapping,
  control_guide_regex = "Non-",
  bin_size_prior = NULL
)
}
\arguments{
\item{counts_array}{an array organized in the following dimension:}

\item{gene_mapping}{a data frame mapping of guide names to gene names. requires the following column names: (1) guide, (2) gene.}

\item{control_guide_regex}{a regular expression used to find/match control guides. default is 'Non-'.}

\item{bin_size_prior}{the expected mass in each bin. If NULL, defaults to uniform (e.g. c(0.25, 0.25, 0.25, 0.25)).}
}
\value{
a water bear object that inference can be performed on.
}
\description{
this function takes a count table and converts it to a water bear object
}
