% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/waterbear.R
\name{wb_run_parallel}
\alias{wb_run_parallel}
\title{parallel waterbear inference}
\usage{
wb_run_parallel(
  wo,
  cluster,
  n_chains = 4,
  n_burnin = 20000,
  n_samples = 40000,
  seed = 42,
  thin = 1,
  summary = TRUE,
  nimble_params = list()
)
}
\arguments{
\item{wo}{waterbear data object from \code{counts_to_wb()}}

\item{cluster}{cluster object from \code{parallel::makeCluster()}. see example.}

\item{n_chains}{total number of MCMC chains. We recommend at least 4.}

\item{n_burnin}{the number of samples to burn in (discard) at the beginning of each chain. we recommend this be half the total number of MCMC samples.}

\item{n_samples}{the total number of MCMC samples generated from each chain. note, we will only keep the samples after the burn in period.}

\item{seed}{either a single number and consecutive numbers will be chosen for each chain or a unique length vector of seeds for each chain.}

\item{nimble_params}{additional parameters to be passed to nimble NIMBLE.}
}
\value{
a set of MCMC samples.
}
\description{
run waterbear in parallel
TODO: add some extraction functions
TODO: add example
}
